+++ rust/candid_parser/tests/assets/ok/example.js
@@ -5,30 +5,8 @@ export const idlFactory = ({ IDL }) => {
   const stream = IDL.Rec();
   const t = IDL.Rec();
   const tree = IDL.Rec();
-  const b = IDL.Tuple(IDL.Int, IDL.Nat);
   const node = IDL.Record({ 'head' : IDL.Nat, 'tail' : list });
   list.fill(IDL.Opt(node));
-  const A = B;
-  B.fill(IDL.Opt(A));
-  tree.fill(
-    IDL.Variant({
-      'branch' : IDL.Record({ 'val' : IDL.Int, 'left' : tree, 'right' : tree }),
-      'leaf' : IDL.Int,
-    })
-  );
-  stream.fill(
-    IDL.Opt(
-      IDL.Record({
-        'head' : IDL.Nat,
-        'next' : IDL.Func([], [stream], ['query']),
-      })
-    )
-  );
-  const s = IDL.Service({
-    'f' : t,
-    'g' : IDL.Func([list], [B, tree, stream], []),
-  });
-  t.fill(IDL.Func([s], [], []));
   const my_type = IDL.Principal;
   List.fill(IDL.Opt(IDL.Record({ 'head' : IDL.Int, 'tail' : List })));
   const nested = IDL.Record({
@@ -73,16 +51,42 @@ export const idlFactory = ({ IDL }) => {
       [IDL.Opt(List), res],
       [],
     );
+  const b = IDL.Tuple(IDL.Int, IDL.Nat);
   const a = IDL.Variant({ 'a' : IDL.Null, 'b' : b });
-  return IDL.Service({
-    'bbbbb' : IDL.Func([b], [], []),
+  const nested_records = IDL.Record({
+    'nested' : IDL.Opt(IDL.Record({ 'nested_field' : IDL.Text })),
+  });
+  const my_variant = IDL.Variant({
+    'a' : IDL.Record({ 'b' : IDL.Text }),
+    'c' : IDL.Opt(IDL.Record({ 'd' : IDL.Text })),
+  });
+  const A = B;
+  B.fill(IDL.Opt(A));
+  tree.fill(
+    IDL.Variant({
+      'branch' : IDL.Record({ 'val' : IDL.Int, 'left' : tree, 'right' : tree }),
+      'leaf' : IDL.Int,
+    })
+  );
+  stream.fill(
+    IDL.Opt(
+      IDL.Record({
+        'head' : IDL.Nat,
+        'next' : IDL.Func([], [stream], ['query']),
+      })
+    )
+  );
+  const s = IDL.Service({
     'f' : t,
+    'g' : IDL.Func([list], [B, tree, stream], []),
+  });
+  t.fill(IDL.Func([s], [], []));
+  return IDL.Service({
     'f1' : IDL.Func(
         [list, IDL.Vec(IDL.Nat8), IDL.Opt(IDL.Bool)],
         [],
         ['oneway'],
       ),
-    'g' : IDL.Func([list], [B, tree, stream], []),
     'g1' : IDL.Func(
         [my_type, List, IDL.Opt(List), nested],
         [IDL.Int, broker, nested_res],
@@ -110,6 +114,14 @@ export const idlFactory = ({ IDL }) => {
         ],
         ['composite_query'],
       ),
+    'y' : IDL.Func(
+        [nested_records],
+        [IDL.Tuple(nested_records, my_variant)],
+        ['query'],
+      ),
+    'f' : t,
+    'g' : IDL.Func([list], [B, tree, stream], []),
+    'bbbbb' : IDL.Func([b], [], []),
   });
 };
 export const init = ({ IDL }) => { return []; };